+++ pkg/data/dataset_test.go
@@ -6,7 +6,7 @@ import (
 )
 
 func TestDataset(t *testing.T) {
-	dataset := NewNumberDataset([]byte{150, 100}, []byte{24, 25})
+	dataset := NewNumberDataset([]float64{150, 100}, []float64{24, 25})
 	for sample := range dataset.Iter(2) {
 		fmt.Print(sample)
 	}
